# golang/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk manajemen penjadwalan sekolah dengan Guided GA-GWO algorithm",
        "title": "Timetabling Backend API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:12000",
    "basePath": "/api",
    "paths": {
...
        "/kelas": {
            "get": {
                "description": "Get a list of all kelas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Get all kelas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Kelas"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new kelas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Create kelas",
                "parameters": [
                    {
                        "description": "Kelas data",
                        "name": "kelas",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Kelas"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Kelas"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/kelas/{id}": {
            "get": {
                "description": "Get a specific kelas by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Get kelas by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Kelas ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Kelas"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing kelas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Update kelas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Kelas ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated kelas data",
                        "name": "kelas",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Kelas"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Kelas"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a kelas by ID",
                "tags": [
                    "Kelas"
                ],
                "summary": "Delete kelas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Kelas ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mapel": {
            "get": {
                "description": "Get a list of all mapel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Get all mapel",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Mapel"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new mapel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Create mapel",
                "parameters": [
                    {
                        "description": "Mapel data",
                        "name": "mapel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Mapel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Mapel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mapel/{id}": {
            "get": {
                "description": "Get a specific mapel by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Get mapel by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mapel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Mapel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing mapel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Update mapel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mapel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated mapel data",
                        "name": "mapel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Mapel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Mapel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a mapel by ID",
                "tags": [
                    "Mapel"
                ],
                "summary": "Delete mapel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mapel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Kelas": {
            "type": "object",
            "properties": {
                "kelas_id": {
                    "type": "integer"
                },
                "nama_kelas": {
                    "type": "string"
                },
                "tingkatan": {
                    "type": "integer"
                }
            }
        },
        "models.Mapel": {
            "type": "object",
            "properties": {
                "jam_per_minggu": {
                    "type": "integer"
                },
                "kode_mapel": {
                    "type": "string"
                },
                "mapel_id": {
                    "type": "integer"
                },
                "mgmp": {
                    "type": "string"
                },
                "nama_mapel": {
                    "type": "string"
                }
            }
        },
...
    }
}